Add Task.java, ToDoList.java, Format.java SirDuke.java: Move printing of commands to ToDoList class, add switch statement for command execution
Each task is as string.

Strings cannot encapsulate whether the task was completed or not.

Let's create a Task class, Task.java to encapsulate the task name and completion status.
Let's create a method to get the name of the Task as a String.

This will help us add other properties of Tasks in future versions (like deadlines)

There is no data structure to store Tasks.

Tasks need to be stored and list needs to be displayed.

Let's create a ToDoList class, ToDoList.java, that contains a HashMap(with variable name toDoList) with task index as the key and the Task as the value.
Let's create a method to add a Task to the list when a String is provided.
Let's create a method to display the index and name of all Tasks in the Hashmap.

This allows us to add other functionalities of the ToDoList in future versions (like deletion of tasks)

SirDuke class prints all commands received.

SirDuke class receives, executes and implements the execution of each command, and this complicates
the function of the SirDuke class.

Let's create a switch statement to call the appropriate method based on the given command.
Let's remove unnceccessary print statements.

We create a switch statement to make it easy to add code for other commands to be handled by the SirDuke class in future versions.
We move implementation of execution of commands to the ToDoList class so that SirDuke class only has to receive command
and decide the corresponding action to take, this categorises the functions of each class and makes it easier to know where to
write code when writing features in future versions.

HORIZONTAL_LINE is common across multiple classes.

We have to create a HORIZONTAL_LINE every time we intend to use it in a class.

Let's create a Format class, Format.java that contains all objects we intend to use for formatting printed output.

This allows us to add other formatting objects and use them across all classes that need them without having to re-define them
every time we need to use them.